pyproject.toml
[project]
name = "agentichr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-groq>=0.3.6",
    "langgraph>=0.5.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
